# chore(utils): Narrow down sysinfo data